Don't abort after AnteHandler if CheckTx
Currently we are aborting running a transaction after the AnteHandler in
case it is a checkTx. This means that the handler never gets the chance
to check the validity of a transaction. Furthermore the AnteHandler
should not handle CheckTx logic.

The AnteHandler should handle global validation, whereas each Handler
should handle module validation.